+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support generic type for datatypes and functions.
 - Support file dump in output directory.
 - Implement lookahead branching heuristic.
+- Support interactive debug mode.
 
 ### Removed